pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pixelanalizer"
version = "2.0.0"
description = "Satellite image color analyzer — built for scientific use"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["satellite", "image", "analysis", "remote-sensing", "color", "hsv"]
dependencies = [
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
pixelanalizer = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
